Actually prevent actions based on sync lock
